+++ apps/web/src/app/(public)/agents/remote-control/[sessionId]/components/RemoteTerminal/components/MobileToolbar/index.ts
@@ -1,0 +1,1 @@
+export { MobileToolbar } from "./MobileToolbar";
