+++ apps/web/src/app/(public)/agents/remote-control/[sessionId]/components/RemoteTerminal/index.ts
@@ -1,2 +1,1 @@
 export { RemoteTerminal } from "./RemoteTerminal";
-export { RemoteTerminalLoader } from "./RemoteTerminalLoader";
