+++ apps/web/src/app/(public)/agents/remote-control/[sessionId]/components/RemoteTerminalLoader/index.ts
@@ -1,0 +1,1 @@
+export { RemoteTerminalLoader } from "./RemoteTerminalLoader";
